[refactor] Divide interface for service  - 타임딜 서비스 open함수 인터페이스 #28
