+++ timeDeal/src/main/java/com/bjcareer/stockservice/RedisConfig.java
@@ -1,12 +1,13 @@
 package com.bjcareer.stockservice;
 
-import org.redisson.config.MasterSlaveServersConfig;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 import org.redisson.Redisson;
 import org.redisson.api.RedissonClient;
 import org.redisson.config.Config;
+import org.redisson.config.SingleServerConfig;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
 import com.bjcareer.stockservice.timeDeal.listener.RedisListenerService;
 import com.bjcareer.stockservice.timeDeal.repository.CouponRepository;
 import com.bjcareer.stockservice.timeDeal.repository.EventRepository;
@@ -15,11 +16,8 @@ import com.bjcareer.stockservice.timeDeal.repository.InMemoryEventRepository;
 @Configuration
 public class RedisConfig {
 
-	@Value("${redis.address.master}")
+	@Value("${redis.address}")
 	private String redisMasterAddress;
-
-	@Value("${redis.address.slave}")
-	private String redisSlaveAddress;
 
 	@Value("${redis.password}")
 	private String redisPassword;
@@ -31,23 +29,12 @@ public class RedisConfig {
 	public RedissonClient redissonClient() {
 		Config config = new Config();
 
-		MasterSlaveServersConfig masterSlaveServersConfig = config.useMasterSlaveServers();
+		SingleServerConfig singleServerConfig = config.useSingleServer();
 
 		// 마스터 주소 설정
-		masterSlaveServersConfig.setMasterAddress(redisMasterAddress);
-		masterSlaveServersConfig.setDatabase(redisDatabase);
-		masterSlaveServersConfig.setPassword(redisPassword);
-
-		// 슬레이브 주소 설정
-		masterSlaveServersConfig.addSlaveAddress(redisSlaveAddress);
-
-		masterSlaveServersConfig.setMasterConnectionMinimumIdleSize(70);
-		masterSlaveServersConfig.setSlaveConnectionMinimumIdleSize(70);
-		masterSlaveServersConfig.setSlaveConnectionPoolSize(70);
-		masterSlaveServersConfig.setMasterConnectionPoolSize(70);
-
-		masterSlaveServersConfig.setKeepAlive(true);
-
+		singleServerConfig.setAddress(redisMasterAddress);
+		singleServerConfig.setDatabase(redisDatabase);
+		singleServerConfig.setPassword(redisPassword);
 
 		return Redisson.create(config);
 	}
