+++ timeDeal/src/main/java/com/bjcareer/stockservice/timeDeal/in/api/dto/TimeDealDTO.java
@@ -1,12 +1,8 @@
-package com.bjcareer.stockservice.timeDeal.controller.dto;
+package com.bjcareer.stockservice.timeDeal.in.api.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.util.UUID;
-
-import com.bjcareer.stockservice.timeDeal.domain.coupon.Coupon;
 
 public class TimeDealDTO {
